Adding register pages fragments or homw

+++ FirstMVCProject/src/main/java/com/Learning/repo/userRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Learning.entities.User;
 
-public interface userRepo extends JpaRepository<User, Integer>{
+public interface userRepo extends JpaRepository<User, Long>{
 
 }
